@@ -4,7 +4,7 @@ version: 1.1.0
 homepage: https://github.com/xrr2016/flutter_tree
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
